Refactor so globbing works with simplecache:: prefix
The whole motivation for this small refactor was that the line
`paths = self.get_fs().glob(self._glob_path)`
was failing when trying to use the simplecache:: prefix for fsspec.

We almost never need the prefix (only when trying to actually download
the file), so I did a small refactor so `self.urlpath` typically refers
to the path without the prefix.

This is hard to test without actually accessing s3 files, I'm afraid.